Mitocarta: an inventory of mammalian mitochondrial genes

MitoCarta2.0 is an inventory of 1158 human genes encoding proteins with strong support of mitochondrial localization. To generate this inventory, the Mootha lab at MGH performed mass spectrometry of mitochondria isolated from fourteen tissues, assessed protein localization through large-scale GFP tagging/microscopy, and integrated these results with six other genome-scale datasets of mitochondrial localization, using a Bayesian approach. MitoCarta2.0, released 2015, is an update from the MitoCarta1.0 inventory published in Cell, 2008.

data(mitocarta)

Format

A data frame with 19,244 rows and 43 variables:

training_dataset

Tmito, Tnot_mito, or T_possible_mito (indicating evidence based on NCBI GO mitochondrial annotation or MitoP2 database, but not in Tmito)

human_gene_id

NCBI Entrez Gene ID for human ortholog of mouse gene (based on reciprocal BLASTP hit, expect<1e-3)

mouse_ortholog_gene_id

NCBI Entrez Gene ID for mouse gene

symbol

Official Entrez gene symbol

synonyms

NCBI Entrez gene synonyms (UniProt added at end)

description

NCBI Entrez gene description

ensembl_gene_id

Ensembl Gene Identifier (based on Ensembl mapper)

protein_length

Length of longest RefSeq protein isoform

target_p_score

TargetP confidence score (1-5, 1 is most confident) of mitochondrial targeting signal. Score 0 indicates no mitochondrial prediction

mito_domain_score

MitoDomain indicates exclusively mito domain; SharedDomain indicates shared mito & non-mito domain; NonMitoDomain indicates exclusively non-mito domain; NA indicates unknown

coexpression_gnf_n50_score

N50 score for coexpression across mouse GNFv3 tissue atlas (N50 = number of 50 nearest neighbors that are in Tmito)

pgc_induction_score

Foldchange in transcript induction following overexpression of PGC1a, known to induce mitochondrial biogenesis

yeast_mito_homolog_score

OrthologMitoHighConf (HomologMitoHighConf) if yeast ortholog (homolog) is annotated mitochondrial in SGD with high confidence; OrthologMitoLowConf (HomologMitoLowConf) if yeast ortholog (homolog) is annotated mitochondrial in SGD with low confidence (dual localized or only based on high throughput data), NoMitoHomolog indicates no yeast homolog or homolog is not annotated mitochondrial in SGD

rickettsia_homolog_score

Ortholog if 1:1 ortholog between mouse and Rickettsia prowazekii; Homolog if homology between mouse and Rickettsia prowazekii, otherwise NoHomolog

msms_score

Category indicating MS/MS abundance (coverage) and enrichment in subtractive proteomics (either pure-enriched, crude-enriched, or ambiguous)

mcarta2_score

Naïve Bayes score (sum of log-likelihood ratios for each of 7 above features)

mcarta2_fdr

Estimated corrected false discovery rate for all predictions >= given score

mcarta2_list

1 if gene is on final MitoCarta2.0 list

mcarta2_evidence

Type of experimental support

hg19_chromosome

Chromosome (hg19 assembly)

hg19_start

Start position (hg19)

hg19_stop

Stop position (hg19)

msms_num_tissues

Number 0-14 tissues where gene products were found by MS/MS

msms_num_peptides_unique

Number of unique peptides, based on pooling spectra from 14 tissues

msms_num_spectra

Number of spectra corresponding to this gene, based on pooling spectra from 14 tissues

msms_total_intensity

total peak intensity, based on pooling spectra from 14 tissues

msms_percent_coverage

Coverage (percent of amino acids covered by MS/MS spectra), based on pooling spectra from 14 tissues

tissues

List of tissues with evidence of protein, based on 14 tissues

cerebrum_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

cerebellum_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

brainstem_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

spinalcord_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

kidney_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

liver_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

heart_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

skeletalmuscle_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

adipose_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

smallintestine_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

largeintestine_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

stomach_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

placenta_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

testis_total_peak_intensity_log10

Total peak intensity of spectra found in this tissue (log base 10)

hpa_primary_subcellular_localization_2015

Human Protein Atlas [September 2015] designation of primary subcellular localization based on immunofluorescence of antibodies in 3 human cell lines; Text in parenthesis indicates antibody reliability (where APE=antibody protein expression)

Source

https://www.broadinstitute.org/scientific-community/science/programs/metabolic-disease-program/publications/mitocarta/mitocarta-in-0